+++ README
@@ -1,10 +1,11 @@
-소방 방재 프로그램
+#소방 방재 프로그램
+<div>
+    NodeMCU보드에 각종 센서로부터 데이터값을 받아옵니다.\n
+    받아온 raw data를 MQTT를 이용해 내부 네트워크에 존재하는 브로커(rpi)에 publish(데이터 전송) 하여\n
+    같은 네트워크의 브로커와 topic을 구독하여 pub한 raw 데이터를 subscribe 하여 받아온 값을 maria DB에 저장합니다.\n
+    Flask를 이용하여 웹서버에 DB에 쌓이는 데이터를 실시간으로 그래프를 갱신시켜 웹으로 나타냅니다.\n
+</div>
 
-
-NodeMCU보드에 각종 센서로부터 데이터값을 받아옵니다.
-받아온 raw data를 MQTT를 이용해 내부 네트워크에 존재하는 브로커(rpi)에 publish(데이터 전송) 하여
-같은 네트워크의 브로커와 topic을 구독하여 pub한 raw 데이터를 subscribe 하여 받아온 값을 maria DB에 저장합니다.
-Flask를 이용하여 웹서버에 DB에 쌓이는 데이터를 실시간으로 그래프를 갱신시켜 웹으로 나타냅니다.
 
 pub.c는 센서 데이터 수집 및 전송 역할을 한다.
 subbroker(topic)_v1.x 는 브로커 및 subscribe(수집)하는 역할을 한다.
@@ -12,5 +13,6 @@ makedata.py는 더미데이터를 생성하여 학습하는데 필요한 가상 
 
 시계열 데이터로 생성된 데이터를 전처리하여 LSTM을 이용하여 예측값을 도출할 수 있다.
 
-사용 기술
-통신 프로토콜 : MQTT
+Tech
+Flask, Python, js, html, json
+MQTT
