+++ sensordata_web/web.py
@@ -1,5 +1,4 @@
 from flask import Flask, render_template, request
-#import sqlite3
 import mariadb
 import sys
 import json
